renamed sample package to Buzzword
