+++ .idea/vcs.xml
@@ -3,6 +3,5 @@
   <component name="VcsDirectoryMappings">
     <mapping directory="" vcs="Git" />
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
-    <mapping directory="C:/repos" vcs="Git" />
   </component>
 </project>
