+++ src/Buzzword/AppMessageDialogSingleton.java
@@ -1,4 +1,4 @@
-package sample;
+package Buzzword;
 
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
