+++ src/Buzzword/Main.java
@@ -1,4 +1,4 @@
-package sample;
+package Buzzword;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
