+++ src/Buzzword/YesNoCancelDialogSingleton.java
@@ -1,4 +1,4 @@
-package sample;
+package Buzzword;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
